# apps/frontend/resource/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/students/GetStudents": {
      "get": {
        "operationId": "StudentController_getAllStudents",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Students"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/students/GetStudentsByCourseWithAttendances/{id}": {
      "get": {
        "operationId": "StudentController_getStudentsByCourseWithAttendances",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentAttendanceDto"
                }
              }
            }
          }
        },
        "tags": [
          "Students"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/students/GetStudentsByCourseWithPayments/{id}": {
      "get": {
        "operationId": "StudentController_getStudentsByCourseWithPayments",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentsDto"
                }
              }
            }
          }
        },
        "tags": [
          "Students"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/students/UpdateAttendances": {
      "put": {
        "operationId": "StudentController_updateAttendance",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Students"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/students/UpdateStudentDetails": {
      "put": {
        "operationId": "StudentController_updateStudentDetails",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentDetailsDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDetailsDTO"
                }
              }
            }
          }
        },
        "tags": [
          "Students"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/locations/GetLocations": {
      "get": {
        "operationId": "LocationsController_getLocations",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LocationDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Locations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/locations/UpdateLocations": {
      "put": {
        "operationId": "LocationsController_updateGroups",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/UpdateLocationsInputDto"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LocationDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Locations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/groups/GetGroups": {
      "get": {
        "operationId": "GroupController_getGroups",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/groups/UpdateGroups": {
      "put": {
        "operationId": "GroupController_updateGroups",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/UpdateGroupsInputDto"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/packages/GetPackages": {
      "get": {
        "operationId": "PackageController_GetPackages",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PackageDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Package"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/packages/CreatePackages": {
      "post": {
        "operationId": "PackageController_CreatePackages",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CreatePackageDto"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PackageDto"
                  }
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PackageDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Package"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/packages/GetPackagesAndCoursesByLocGroupType": {
      "get": {
        "operationId": "PackageController_getPackagesAndCourses",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locationId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageCourseAssignDto"
                }
              }
            }
          }
        },
        "tags": [
          "Package"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/packages/AssignPackagesToCourses": {
      "post": {
        "operationId": "PackageController_assignCoursesToPackages",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AssignPackageToCourseDto"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Package"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Etalon-organization",
    "description": "API documentation for Etalon organization",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AttendanceDto": {
        "type": "object",
        "properties": {
          "lessonDateId": {
            "type": "number"
          },
          "date": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "attended": {
            "type": "boolean"
          }
        },
        "required": [
          "lessonDateId",
          "date",
          "attended"
        ]
      },
      "StudentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "sapId": {
            "type": "number"
          },
          "subdate": {
            "format": "date-time",
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "billCompany": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "zip": {
            "type": "number"
          },
          "address": {
            "type": "string"
          },
          "coupon": {
            "type": "string"
          },
          "vatNum": {
            "type": "string"
          },
          "billingAddressTypeId": {
            "type": "number"
          },
          "childName": {
            "type": "string"
          },
          "childMail": {
            "type": "string"
          },
          "childGrade": {
            "type": "number"
          },
          "childTAJ": {
            "type": "string"
          },
          "specialDiet": {
            "type": "boolean"
          },
          "specialDietDesc": {
            "type": "string"
          },
          "mobile": {
            "type": "string"
          },
          "packageType": {
            "type": "string"
          },
          "packageCode": {
            "type": "string"
          },
          "disease": {
            "type": "boolean"
          },
          "diseaseDesc": {
            "type": "string"
          },
          "discount": {
            "type": "string"
          },
          "discount2": {
            "type": "string"
          },
          "attendance": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttendanceDto"
            }
          }
        },
        "required": [
          "id",
          "sapId",
          "subdate",
          "email",
          "firstname",
          "lastname",
          "billCompany",
          "city",
          "zip",
          "address",
          "coupon",
          "vatNum",
          "billingAddressTypeId",
          "childName",
          "childMail",
          "childGrade",
          "childTAJ",
          "specialDiet",
          "specialDietDesc",
          "mobile",
          "packageType",
          "packageCode",
          "disease",
          "diseaseDesc",
          "discount",
          "discount2",
          "attendance"
        ]
      },
      "StudentAttendanceDto": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "string"
          },
          "students": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudentDto"
            }
          }
        },
        "required": [
          "courseId",
          "students"
        ]
      },
      "PaymentDto": {
        "type": "object",
        "properties": {
          "invoiceDateId": {
            "type": "number"
          },
          "date": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "billerId": {
            "type": "number"
          },
          "payed": {
            "type": "boolean"
          },
          "amount": {
            "type": "number"
          },
          "invoiceNumber": {
            "type": "string"
          }
        },
        "required": [
          "invoiceDateId",
          "date",
          "billerId",
          "payed",
          "amount",
          "invoiceNumber"
        ]
      },
      "StudentPaymentDto": {
        "type": "object",
        "properties": {
          "studentName": {
            "type": "string"
          },
          "Payments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentDto"
            }
          }
        },
        "required": [
          "studentName",
          "Payments"
        ]
      },
      "PaymentsDto": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "string"
          },
          "payments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudentPaymentDto"
            }
          }
        },
        "required": [
          "courseId",
          "payments"
        ]
      },
      "UpdateStudentDetailsDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "sapId": {
            "type": "number"
          },
          "email": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "billCompany": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "zip": {
            "type": "number"
          },
          "address": {
            "type": "string"
          },
          "coupon": {
            "type": "string"
          },
          "vatNum": {
            "type": "string"
          },
          "billingAddressTypeId": {
            "type": "number"
          },
          "childName": {
            "type": "string"
          },
          "childMail": {
            "type": "string"
          },
          "childGrade": {
            "type": "number"
          },
          "childTAJ": {
            "type": "string"
          },
          "specialDiet": {
            "type": "boolean"
          },
          "specialDietDesc": {
            "type": "string"
          },
          "mobile": {
            "type": "string"
          },
          "packageType": {
            "type": "string"
          },
          "packageCode": {
            "type": "string"
          },
          "disease": {
            "type": "boolean"
          },
          "diseaseDesc": {
            "type": "string"
          },
          "discount": {
            "type": "string"
          },
          "discount2": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "sapId",
          "email",
          "firstname",
          "lastname",
          "billingAddressTypeId",
          "childName",
          "specialDiet",
          "disease"
        ]
      },
      "StudentDetailsDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "sapId": {
            "type": "number"
          },
          "subdate": {
            "format": "date-time",
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "billCompany": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "zip": {
            "type": "number"
          },
          "address": {
            "type": "string"
          },
          "coupon": {
            "type": "string"
          },
          "vatNum": {
            "type": "string"
          },
          "billingAddressTypeId": {
            "type": "number"
          },
          "childName": {
            "type": "string"
          },
          "childMail": {
            "type": "string"
          },
          "childGrade": {
            "type": "number"
          },
          "childTAJ": {
            "type": "string"
          },
          "specialDiet": {
            "type": "boolean"
          },
          "specialDietDesc": {
            "type": "string"
          },
          "mobile": {
            "type": "string"
          },
          "packageType": {
            "type": "string"
          },
          "packageCode": {
            "type": "string"
          },
          "disease": {
            "type": "boolean"
          },
          "diseaseDesc": {
            "type": "string"
          },
          "discount": {
            "type": "string"
          },
          "discount2": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "sapId",
          "subdate",
          "email",
          "firstname",
          "lastname",
          "billCompany",
          "city",
          "zip",
          "address",
          "coupon",
          "vatNum",
          "billingAddressTypeId",
          "childName",
          "childMail",
          "childGrade",
          "childTAJ",
          "specialDiet",
          "specialDietDesc",
          "mobile",
          "packageType",
          "packageCode",
          "disease",
          "diseaseDesc",
          "discount",
          "discount2"
        ]
      },
...
      "LocationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "description"
        ]
      },
      "UpdateLocationsInputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "nullable": true
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "description"
        ]
      },
      "GroupDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "description"
        ]
      },
      "UpdateGroupsInputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "nullable": true
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "description"
        ]
      },
...
      "PackageDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "id": {
            "type": "number"
          },
          "packageId": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "locationId": {
            "type": "number"
          },
          "locationDesc": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "groupId": {
            "type": "number"
          },
          "groupDesc": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "id",
          "packageId",
          "price",
          "locationId",
          "locationDesc",
          "active",
          "groupId",
          "groupDesc"
        ]
      },
      "CreatePackageDto": {
        "type": "object",
        "properties": {
          "packageId": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "type": {
            "type": "string"
          },
          "locationId": {
            "type": "number"
          },
          "active": {
            "type": "boolean"
          },
          "groupId": {
            "type": "number"
          }
        },
        "required": [
          "packageId",
          "price",
          "type",
          "locationId",
          "active",
          "groupId"
        ]
      },
      "CourseAssignDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "courseId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "courseId"
        ]
      },
      "PackageAssignDto": {
        "type": "object",
        "properties": {
          "packageId": {
            "type": "string"
          }
        },
        "required": [
          "packageId"
        ]
      },
      "AssingDto": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "number"
          },
          "packageId": {
            "type": "string"
          }
        },
        "required": [
          "courseId",
          "packageId"
        ]
      },
      "PackageCourseAssignDto": {
        "type": "object",
        "properties": {
          "courses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CourseAssignDto"
            }
          },
          "packages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PackageAssignDto"
            }
          },
          "assignments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssingDto"
            }
          }
        },
        "required": [
          "courses",
          "packages",
          "assignments"
        ]
      },
      "AssignPackageToCourseDto": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "number"
          },
          "packageId": {
            "type": "string"
          },
          "assign": {
            "type": "boolean"
          }
        },
        "required": [
          "courseId",
          "packageId",
          "assign"
        ]
      },
...
    }
  }
}